Allow destroying extra host capabilities
Though there was DB functionality already written for this, it was
never exposed at the API layer. This prevented hosts from having
extra capabilities explicitly removed.

To remove an extra capability, specify the capability name and
set the value explicitly to `None` (when sending JSON, send `null`.)

Existing restrictions apply; it is not possible to remove a
capability in use explicitly by an active or pending reservation.

Closes-Bug: #1674524
Change-Id: Ic0a555b6f8cd23d213189c1420781dd4e6d4e848
Signed-off-by: Matt Crees <mattc@stackhpc.com>
Signed-off-by: Pierre Riteau <pierre@stackhpc.com>