WIP: Option to use flavor_id to request reservation
Instead of creating custom RAM and CPU ratios, we offer the option
to reserve a specific number of a pre-defined flavor.

When you specify the flavor, we still allow amount and affinity,
but we ignore resource_properties, vcpus, memory_mb and disk_gb,
as these are all taken from the flavor instead.

TODO: should we have the option to block non-flavor based instance
reservation via a configuration option?

Change-Id: I9803a7ee21011584263c83fb982d403fa605ae34