Fix up read session handling
Firstly, we raise exceptions during the database tear down
which makes unit tests fail when the "database is locked"
error is generated.

The first test failing was model_query which was executing
a query outside of the facade context handler. This meant
the session wasn't being closed, which blocks future writes
when using sqlite.

This was spotted, because it was causing issues with the DB
tests for the blazar flavor plugin, when a bad read was
later followed by a write that failed due to the read
that was still holding the lock.

Finally we fix up all the other failing unit tests,
where similiarly there were read queries being made
outside the scope of the context manager that created
the read session.

Change-Id: Ifec8cc37e029656537a080245269e0d3b5ccbe2c