Add extra on_end calls on reservation failure
If we use check_create to keep state about reserverations
we need an on_end callback to be attempted when the lease
is deleted early due to a failure, such as reserve_resource
failing due to some race with another reservation request.

In other error cases, we would expect a lease db object
to still be created at the end of the create lease call,
such that delete will be called that triggers on_end.

One alternative is to move check_create after operations
such as reserve_resource, but in that case we need to
make all of the plugins capable of cleaning up the resources
they have just created, which seems a bad idea when it is
not uncommon that a request is rejected due to the
enforcement logic.

Change-Id: If4ff1c8da32c3be47bf1523058f83effd7340f02